+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - connection test command
 - store author name for version
 
+## Changed
+- Nextcloud v17 is no longer supported
+
 ## 6.0.2
 ## Changed
 - compatible with Nextcloud 20
@@ -21,7 +24,8 @@
 ## 5.0.0
 ## Added
 - support for OpenDocument Templates
-- Japanese translation- certificate verification setting
+- Japanese translation
+- certificate verification setting
 - version history
 
 ## Changed