+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+## 6.1.0
 ## Added
 - use guest name from talk
 - connection test command
@@ -8,6 +9,7 @@
 - Italian translation
 
 ## Changed
+- display local time in history
 - Nextcloud v17 is no longer supported
 
 ## 6.0.2